requirements.txt
python-telegram-bot==21.7
python-dotenv==1.0.1
tzdata==2025.2
httpx==0.28.1
